requirements_stream.txt
requests
kafka-python==2.0.2
apache-airflow ==2.10.0
six
apache-airflow-providers-docker